chore: rename getVideoUrl
