+++ packages/utils/functions/getPublicationMediaUrl.ts
@@ -2,12 +2,6 @@ import { STATIC_ASSETS } from '../constants'
 import type { LenstubePublication } from '../custom-types'
 import sanitizeIpfsUrl from './sanitizeIpfsUrl'
 
-export const getVideoUrl = (video: LenstubePublication) => {
-  const url = video?.metadata?.media[0]?.original.url
-  if (!url) return `${STATIC_ASSETS}/images/fallbackThumbnail.png`
-  return sanitizeIpfsUrl(url)
-}
-
 export const getPublicationMediaUrl = (video: LenstubePublication) => {
   const url = video?.metadata?.media[0]?.original.url
   if (!url) return `${STATIC_ASSETS}/images/fallbackThumbnail.png`
